ref: automate `OrderedData` data valuation algorithm
Each data point used to required an optional `Extractor` (now called a
`Valuator`) which accepted a field value and returned the innermost
value as determined by the `Order`.

Now, all that is necessary is to provide which keys may need to be
valuated (i.e. those which are `object`s) and the data ordering will
happen automatically.