fix(ContactTable): attempt to prevent deleting assigned addresses
Apparently when deleting a foreign row that has been assigned to a row
in a table with a column having a nullable foreign key, it will not
check the constraints on that table to see if the deletion will result
in a valid row on that table.

I don't know if there is anything which I can do other than patch this
on the backend, since it should be checking the constraints on tables
which are being changed. But this is not the case.